Fix checkout queries
